requirements.txt
boto3==1.35.36
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5

